Fix bug in average_predictions function
This function creates a list of predictions corresponding to a
list of ids. However it is designed to skip the predictions
containing infs or NaNs. Thus, when a prediction is skipped a
shift between the predictions and the ids is generated.
average_predictions_ids fixes this by returning a list of the
ids that were not skipped.