Fix bug in the validation of models based on embeddings
The data loaders used for validation were using the parameter
`number_of_frames` as the hop size between adjacent patches.
This is only correct for spectrogram-based models. For embedding-based
models the hop size should be 1, as every column in the embedding
matrix corresponds to `number_of_frames` timestamps (1 to 3 seconds
depending on the model).
This fix sets up a variable `xInput` with the correct hop size
depending on the case just as in the train script.