Reimplement preprocess and preprocess_crosseval
preprocess.py is reimplemented to support the new feature extractors.
It is designed rely on the standard Essentia functions by default, and it only imports
TensorFlow-related algorithms when embeddings are required.

preprocess_crosseval.py is modify to rely mostly in preprocess.py methods. It's basically
a wrapper that works with arguments instead of a configuration file (which is useful for
external datasets). This affects where the script writes the index.tsv file: in `data_dir`,
 instead of `datadir/metadata` as before.